Use GenomicSQLite for VCF records and Variants (#2)
Use GenomicSQLite instead of Spark SQL to store and range-join the [g]VCF records and discovered Variants. This improves Spark app efficiency & reliability, by reducing (i) materialized copying of the VCF records and (ii) spikes in the size of the in-memory working set. (We're still leveraging Spark extensively, to parallelize everything and perform final assembly of the pVCF output records.)